% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2latex.R
\name{matrix2latex}
\alias{matrix2latex}
\title{This function print the latex code for a input matrix}
\usage{
matrix2latex(matr, environment.matrix = "bmatrix")
}
\arguments{
\item{matr}{input matrix}
}
\description{
This function print the latex code for a input matrix
}
\examples{
\dontrun{
print.figure('figure.tex'),filename = "yourfigure.pdf",caption = "figure title",label="yourlabel",landscape=FALSE)
}
}

