% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFigure.R
\name{print.figure}
\alias{print.figure}
\title{This function write a tex file to insert the figure into your latex file}
\usage{
print.figure(writeto, filename, placement = "hppt", caption, label, shcap,
  scale = 1, height = FALSE, landscape = FALSE)
}
\arguments{
\item{writeto}{the resulting tex file, full name ended with .tex, a string}

\item{filename}{the filename for the figure, full name, a string}

\item{caption}{figure caption name, a string}

\item{label}{figure label, a string}

\item{landscape}{TRUE or FALSE, landscape or not}
}
\description{
This function write a tex file to insert the figure into your latex file
}
\examples{
\dontrun{
print.figure('figure.tex'),filename = "yourfigure.pdf",caption = "figure title",label="yourlabel",landscape=FALSE)
}
}

