% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_tex.R
\name{tutorial_tex}
\alias{tutorial_tex}
\title{This function write a tex file with certain template}
\usage{
tutorial_tex(writeto, type = "tutorial")
}
\arguments{
\item{writeto}{the resulting tex file, full name ended with .tex, a string}
}
\description{
This function write a tex file with certain template
}
\examples{
\dontrun{
tutorial_tex(writeto='tutorial_temp.tex',type="tutorial")
}
}

